Try and fix + clean up convert_tag_format.py
